moved openFilter into utils file
