+++ client/test/app/queue/components/modalUtils.js
@@ -1,5 +1,5 @@
 import userEvent from '@testing-library/user-event';
-import { screen } from '@testing-library/react';
+import { screen, fireEvent, waitFor } from '@testing-library/react';
 import * as uiActions from 'app/queue/uiReducer/uiActions';
 
 /**
@@ -127,4 +127,15 @@ export const createSpyRequestPatch = (postData) => {
         }
       }
     )));
+};
+
+export const openFilter = async (container) => {
+  const svg = container.querySelectorAll('svg');
+
+  const filter = svg[svg.length - 1];
+
+  fireEvent.click(filter);
+  await waitFor(() => {
+    expect(screen.getByText('Date filter parameters')).toBeInTheDocument();
+  });
 };
